fix(typo): rename separeted_lines to separated_lines
